requirements.txt
# requirements.txt
streamlit
requests
python-dotenv
speechrecognition
gTTS
fpdf
pyaudio
folium
geocoder
plotly